Modularize PostgresStore class
