+++ libs/store-postgres/src/modules/utils.ts
@@ -1,0 +1,33 @@
+/**
+ * Validates the provided namespace.
+ * @param namespace The namespace to validate.
+ * @throws {Error} If the namespace is invalid.
+ */
+export function validateNamespace(namespace: string[]): void {
+  if (namespace.length === 0) {
+    throw new Error("Namespace cannot be empty.");
+  }
+  for (const label of namespace) {
+    if (typeof label !== "string") {
+      throw new Error(
+        `Invalid namespace label '${label}' found in ${namespace}. Namespace labels ` +
+          `must be strings, but got ${typeof label}.`
+      );
+    }
+    if (label.includes(".")) {
+      throw new Error(
+        `Invalid namespace label '${label}' found in ${namespace}. Namespace labels cannot contain periods ('.').`
+      );
+    }
+    if (label === "") {
+      throw new Error(
+        `Namespace labels cannot be empty strings. Got ${label} in ${namespace}`
+      );
+    }
+  }
+  if (namespace[0] === "langgraph") {
+    throw new Error(
+      `Root label for namespace cannot be "langgraph". Got: ${namespace}`
+    );
+  }
+}
